Create reusable BaseSettings class
Extracting parts of the quixstreams.kafka.configuration.ConnectionConfig
to a separate base class so that we can reuse it in the future for the
Schema Registry configuration.

Additionally fixed missing typing.
